feat: Add Edit Screen
Click + Fab takes user to an Edit Screen with a blank Reminder.  Filling out fields and hitting Save Fab returns user to List Screen with the reminder added.  Clicking on a List Item navigates to Edit Screen with info loaded in.  Saving will update that item in list instead of adding a new one.

Note:  This is all done by a list cached in the VM.  Next step is to add a Database and store reminders in the database.