feat: Added Firebase Authentication feat: Started Google Maps integration
User is now presented with a login screen using FirebaseUI and will be prompted to either create an account or login to an existing account. Singout button added.  Still need to determine if I will attempt to backup User reminders in FireStore database.

MapView has been added to the Edit Fragment and necessary code is there to show and allow user to explore the map.  Need to add the logic to get select location and create geofence.